migrated all components to Home component
