+++ src/Components/Resume.jsx
@@ -1,0 +1,8 @@
+export default function Resume(){
+    return(
+        <div>
+            Resume
+        </div>
+    )
+
+}
